Add files via upload
Python script that takes in a fasta file, a PDCUB score file, and a CDS window size W in nucleotides. The script computes tAI for the first W nucleotides and the second W nucleotides. There are two outputs: A scatterplot of second-window tAI against first-window tAI for each transcript, and a scatterplot of first-plus-second-window tAI against PDCUB score for each transcript.